+++ bulletin_board/server/db/schema.rb
@@ -2,15 +2,15 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_03_03_141159) do
+ActiveRecord::Schema.define(version: 2021_04_13_125626) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -33,7 +33,14 @@ ActiveRecord::Schema.define(version: 2021_03_03_141159) do
     t.bigint "byte_size", null: false
     t.string "checksum", null: false
     t.datetime "created_at", null: false
+    t.string "service_name", null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+  end
+
+  create_table "active_storage_variant_records", force: :cascade do |t|
+    t.bigint "blob_id", null: false
+    t.string "variation_digest", null: false
+    t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
 
   create_table "clients", force: :cascade do |t|
@@ -86,8 +93,10 @@ ActiveRecord::Schema.define(version: 2021_03_03_141159) do
     t.integer "iat", null: false
     t.string "author_unique_id", null: false
     t.string "message_type", null: false
+    t.string "message_subtype"
     t.index ["chained_hash"], name: "index_log_entries_on_chained_hash", unique: true
     t.index ["client_id"], name: "index_log_entries_on_client_id"
+    t.index ["election_id", "message_type", "message_subtype", "author_unique_id"], name: "index_log_entries_on_type_subtype_author"
     t.index ["election_id"], name: "index_log_entries_on_election_id"
     t.index ["iat"], name: "index_log_entries_on_iat"
   end
@@ -105,6 +114,7 @@ ActiveRecord::Schema.define(version: 2021_03_03_141159) do
   end
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
   add_foreign_key "election_trustees", "clients", column: "trustee_id"
   add_foreign_key "election_trustees", "elections"
   add_foreign_key "elections", "clients", column: "authority_id"