+++ bulletin_board/server/spec/factories/models.rb
@@ -32,6 +32,7 @@ FactoryBot.define do
       trustees_plus_keys { Trustee.first(3).zip(Test::PrivateKeys.trustees_private_keys) }
       election_id { generate(:election_id) }
       voting_scheme { :dummy }
+      polling_stations { %w(polling_station_1 polling_station_2) }
     end
 
     title { { en: Faker::Lorem.question } }
@@ -46,7 +47,8 @@ FactoryBot.define do
                                                 message: build(:create_election_message, election_id: election.unique_id,
                                                                                          authority_client: election.authority,
                                                                                          voting_scheme: evaluator.voting_scheme,
-                                                                                         trustees_plus_keys: evaluator.trustees_plus_keys))
+                                                                                         trustees_plus_keys: evaluator.trustees_plus_keys,
+                                                                                         polling_stations: evaluator.polling_stations))
     end
 
     trait :key_ceremony do
@@ -136,6 +138,20 @@ FactoryBot.define do
       bulletin_board { true }
       client { nil }
       signed_data { BulletinBoard.sign(message) }
+    end
+
+    trait :vote do
+      transient do
+        message { build(:vote_message, election: election, voter_id: voter_id) }
+        voter_id { generate(:voter_id) }
+      end
+    end
+
+    trait :in_person_vote do
+      transient do
+        message { build(:in_person_vote_message, election: election, voter_id: voter_id) }
+        voter_id { generate(:voter_id) }
+      end
     end
   end